created the valid movement object
